feat: 이미지 로드 관련 처리를 커스텀 훅으로 분리
